requirements.txt
ipdb
trimesh
numpysocket
scikit-learn
bpy-fake-module-x.x # x.x is blender's version